predictor/requirements.txt
tensorflow
scikit-learn
numpy
boto3
kafka-python
joblib
